[refactor] 02/27 토큰 검증 로직 usecase로 이동
